+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Apply Black formatter [#589](https://github.com/IN-CORE/pyincore/issues/589)
 - Internal client connecting to the IN-CORE services [#609](https://github.com/IN-CORE/pyincore/issues/609)
+- MlEnabledCgeJoplin analysis with documentation [#588](https://github.com/IN-CORE/pyincore/issues/588)
 
 
 ## [1.19.0] - 2024-06-12